Cleanup for nn support
