+++ templates/treegrid.html
@@ -968,21 +968,21 @@ function panDetailDisplay() {
 }
 
 function make_tree_fn(_width, _height) {
-     return d3.tree()
-              .size([_width, _height])
-              .separation(function(a, b) {
-                  // Horizontal separation between tree nodes
-                  if (a.parent != b.parent)
-                      // Nodes are not siblings: double separation
-                      return 2.2;
-                  // For siblings, increase separation linearly with their combined
-                  // text (caption) length
-                  // var aLen = a.data.text.length;
-                  var aLen = a.data.n ? Math.max(a.data.text.length, 11) : a.data.text.length;
-                  // var bLen = b.data.text.length;
-                  var bLen = b.data.n ? Math.max(b.data.text.length, 11) : b.data.text.length;
-                  return Math.max(1.20, 0.115 * (aLen + bLen));
-              });
+   return d3.tree()
+      .size([_width, _height])
+         .separation(function(a, b) {
+            // Horizontal separation between tree nodes
+            if (a.parent != b.parent)
+               // Nodes are not siblings: double separation
+               return 2.2;
+               // For siblings, increase separation linearly with their combined
+               // text (caption) length
+               // var aLen = a.data.text.length;
+               var aLen = a.data.n ? Math.max(a.data.text.length, 11) : a.data.text.length;
+               // var bLen = b.data.text.length;
+               var bLen = b.data.n ? Math.max(b.data.text.length, 11) : b.data.text.length;
+               return Math.max(1.20, 0.115 * (aLen + bLen));
+         });
  }
 {% if nn_parsing_enabled -%}
  // Configure the tree layout object
@@ -990,9 +990,7 @@ var nnTreeData = null;
 var nnRoot = null;
 var nntree_fn = null;
 function get_nntree(text) {
-     payload = {
-         text:text
-     }
+   payload = {text: text};
     $.ajax({
         url: "{{ url_for('routes.nnparse_api') }}",
         data: payload,
